+++ scripts/DBUpdate-to-5.oracle.sql
@@ -1,0 +1,9 @@
+-- ----------------------------------------------------------
+--  driver: oracle
+-- ----------------------------------------------------------
+SET DEFINE OFF;
+SET SQLBLANKLINES ON;
+CREATE INDEX link_relation_list_source ON link_relation (source_object_id, source_key, state_id);
+CREATE INDEX link_relation_list_target ON link_relation (target_object_id, target_key, state_id);
+SET DEFINE OFF;
+SET SQLBLANKLINES ON;